# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  admin: "admin",
  member: "member",
} as const;

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: Role;
  createdAt: string;
}

export interface SignupBody {
  name: string;
  email: string;
  password: string;
  role?: Role;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface ProjectSummary {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  ownerId: number | null;
  /** @nullable */
  ownerName: string | null;
  createdAt: string;
  memberCount: number;
  taskCount: number;
}

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  dueDate: string | null;
  projectId: number;
  /** @nullable */
  projectName: string | null;
  /** @nullable */
  assigneeId: number | null;
  /** @nullable */
  assigneeName: string | null;
  /** @nullable */
  createdBy: number | null;
  /** @nullable */
  createdByName: string | null;
  createdAt: string;
}

export interface ProjectDetail {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  ownerId: number | null;
  /** @nullable */
  ownerName: string | null;
  createdAt: string;
  members: User[];
  tasks: Task[];
}

export interface CreateProjectBody {
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface UpdateProjectBody {
  name?: string;
  /** @nullable */
  description?: string | null;
}

export interface AddMemberBody {
  userId: number;
}

export interface CreateTaskBody {
  title: string;
  /** @nullable */
  description?: string | null;
  projectId: number;
  /** @nullable */
  assigneeId?: number | null;
  priority?: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
}

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: TaskStatus;
  priority?: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  assigneeId?: number | null;
}

export interface TasksByStatus {
  todo: number;
  in_progress: number;
  done: number;
}

export interface Dashboard {
  totalProjects: number;
  totalTasks: number;
  tasksByStatus: TasksByStatus;
  overdueCount: number;
  overdueTasks: Task[];
  recentTasks: Task[];
  myTasks: Task[];
}

export type ListTasksParams = {
  projectId?: number;
  assigneeId?: number;
  status?: ListTasksStatus;
  overdue?: boolean;
};

export type ListTasksStatus =
  (typeof ListTasksStatus)[keyof typeof ListTasksStatus];

export const ListTasksStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;