# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Team Task Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "200":
          description: Signed up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects visible to the current user
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project with members and tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /projects/{id}/members:
    post:
      operationId: addProjectMember
      tags: [projects]
      summary: Add a member to a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberBody"
      responses:
        "200":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /projects/{id}/members/{userId}:
    delete:
      operationId: removeProjectMember
      tags: [projects]
      summary: Remove a member from a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    Role:
      type: string
      enum: [admin, member]
    TaskStatus:
      type: string
      enum: [todo, in_progress, done]
    TaskPriority:
      type: string
      enum: [low, medium, high]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        createdAt:
          type: string
      required: [id, name, email, role, createdAt]
    SignupBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          $ref: "#/components/schemas/Role"
      required: [name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    ProjectSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
        memberCount:
          type: integer
        taskCount:
          type: integer
      required: [id, name, description, ownerId, ownerName, createdAt, memberCount, taskCount]
    ProjectDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
      required: [id, name, description, ownerId, ownerName, createdAt, members, tasks]
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
      required: [name]
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
    AddMemberBody:
      type: object
      properties:
        userId:
          type: integer
      required: [userId]
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        dueDate:
          type: ["string", "null"]
        projectId:
          type: integer
        projectName:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        createdBy:
          type: ["integer", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, title, description, status, priority, dueDate, projectId, projectName, assigneeId, assigneeName, createdBy, createdByName, createdAt]
...