# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Task } from "./task";
import type { User } from "./user";

export interface ProjectDetail {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  ownerId: number | null;
  /** @nullable */
  ownerName: string | null;
  createdAt: string;
  members: User[];
  tasks: Task[];
}