# lib/api-zod/src/generated/types/signupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from "./role";

export interface SignupBody {
  name: string;
  email: string;
  password: string;
  role?: Role;
}