# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from "./taskPriority";
import type { TaskStatus } from "./taskStatus";

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  dueDate: string | null;
  projectId: number;
  /** @nullable */
  projectName: string | null;
  /** @nullable */
  assigneeId: number | null;
  /** @nullable */
  assigneeName: string | null;
  /** @nullable */
  createdBy: number | null;
  /** @nullable */
  createdByName: string | null;
  createdAt: string;
}