# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from "./role";

export interface User {
  id: number;
  name: string;
  email: string;
  role: Role;
  createdAt: string;
}